It turns out the simpler way that just use `json.Unmarshal` is faster than calling `Decode` to parse it manually.

```
name               old time/op    new time/op    delta
JSONKeys/short-12    2.54µs ± 1%    1.60µs ± 2%   ~     (p=0.100 n=3+3)
JSONKeys/long-12     5.61µs ± 0%    3.61µs ± 0%   ~     (p=0.100 n=3+3)

name               old alloc/op   new alloc/op   delta
JSONKeys/short-12    1.64kB ± 0%    1.06kB ± 0%   ~     (p=0.100 n=3+3)
JSONKeys/long-12     2.74kB ± 0%    1.30kB ± 0%   ~     (p=0.100 n=3+3)

name               old allocs/op  new allocs/op  delta
JSONKeys/short-12      45.0 ± 0%      20.0 ± 0%   ~     (p=0.100 n=3+3)
JSONKeys/long-12       89.0 ± 0%      37.0 ± 0%   ~     (p=0.100 n=3+3)
```