@suzaku 2020-08-13
Never mind. It seems `write` depends on the order of `keys` and assumes that "timeField" is the first key.