@phuslu 2020-08-15
Thanks. Yes, iterate on a golang map will get a shuffled/random order.
So jsonKeys was introduced because internal users ask for `fields should keep orderred` feature on `ConsoleWriter`