@suzaku 2020-08-15
I believe we can do better, because parsing keys only takes longer than decoding the while `JSON` just feels wrong. That said, it won't be an easy job unless we use some other `JSON` library.