@phuslu 2020-08-15
Agreed.

To keeps "No Dependencies" feature, I'm looking for a simplified/bundled `JSON` parser like,
  1. pkg/json https://github.com/pkg/json
  2. gjson https://github.com/tidwall/gjson
I'm considering how to implement a "only top level" json parser base on their codes.